feat : ml server

+++ src/main/java/se/ton/t210/dto/ExpectScoreResponse.java
@@ -7,9 +7,9 @@ public class ExpectScoreResponse {
 
     private final int currentScore;
     private final float currentPercentile;
-    private final float expectedPassPercent;
+    private final int expectedPassPercent;
 
-    public ExpectScoreResponse(int currentScore, float currentPercentile, float expectedPassPercent) {
+    public ExpectScoreResponse(int currentScore, float currentPercentile, int expectedPassPercent) {
         this.currentScore = currentScore;
         this.currentPercentile = currentPercentile;
         this.expectedPassPercent = expectedPassPercent;
